+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   - Async save the analysis to cache db to make startup faster.
   - Support Auto refresh settings memoizing with a ttl of 1 second avoiding the need of restarting server when changing configs. #502
   - Bump clj-kondo to `2021.09.26-20211015.212311-11` adding new `gen-class` linter and other fixes/improvements.
+  - Remove unused duplicate require if any. #527
   
 - Editor
   - Add reference code lens for ns forms. #578